feat: implement active PR and command session handling with GitHub CLI integration
